some more changes

+++ Models/userMode.js
@@ -9,22 +9,49 @@ const userModel = {
         return result;
     },
     creatUser: async function (DTO) {
-        userValidator(DTO);
-        const result = await DAO.creatuser(DTO);
-        return result;
+        // validating the data 
+
+        const validateObj = userValidator(DTO);
+        if (validateObj.All) {
+            const result = await DAO.creatuser(DTO);
+            return result;
+        }
+        else {
+            throw new error("Validation fail")
+        }
     },
     updateUser: async function (Data) {
-        userValidator({ ...Data.change, _Id_: Data._Id_ });
+        // validating the data 
+
+        const validationObj = userValidator({ ...Data.change, _Id_: Data._Id_ });
+        for (let key in Data.change) {
+            if (!validatorObj[key]) {
+                throw new error("validation fail")
+            }
+        }
+        if (!validatorObj["_Id_"]) {
+            return new error("Validation fail")
+        }
         const result = await DAO.updateUser(Data);
         return result
     },
     getUser: async function (Data) {
-        userValidator(validateId(Data._Id_));
+        // validating the data 
+
+        const validattionObj = userValidator(validateId(Data._Id_));
+        if (!validattionObj["_Id_"]) {
+            return new error("Validation fail")
+        }
         const result = DAO.getUser(Data._Id_);
         return result
     },
-    deleteUser : async function (Data) {
-        userValidator(validateId(Data._Id_));
+    deleteUser: async function (Data) {
+        // validating the data 
+
+        const validattionObj = userValidator(validateId(Data._Id_));
+        if (!validattionObj["_Id_"]) {
+            return new error("Validation fail")
+        }
         const result = DAO.deleteUser(Data._Id_);
         return result;
     }
